Added clean_iob_prefixes and used it in relevant checks. Added new file token_classification_utils.py and moved utils there
