+++ deepchecks/nlp/checks/model_evaluation/prediction_drift.py
@@ -17,6 +17,7 @@ from deepchecks.core import CheckResult
 from deepchecks.core.errors import DeepchecksValueError
 from deepchecks.nlp import Context, TrainTestCheck
 from deepchecks.nlp.task_type import TaskType
+from deepchecks.nlp.utils.token_classification_utils import clean_iob_prefixes
 from deepchecks.utils.abstracts.prediction_drift import PredictionDriftAbstract
 from deepchecks.utils.distribution.preprocessing import convert_multi_label_to_multi_class
 
@@ -156,8 +157,8 @@ class PredictionDrift(PredictionDriftAbstract, TrainTestCheck):
                           'classification tasks. Using drift_mode="prediction" instead.', UserWarning)
 
         if context.task_type == TaskType.TOKEN_CLASSIFICATION:
-            train_predictions = clean_iob_format(np.concatenate(model.predict(train_dataset)).reshape(-1, 1))
-            test_predictions = clean_iob_format(np.concatenate(model.predict(test_dataset)).reshape(-1, 1))
+            train_prediction = clean_iob_prefixes(np.concatenate(model.predict(train_dataset)).reshape(-1, 1))
+            test_prediction = clean_iob_prefixes(np.concatenate(model.predict(test_dataset)).reshape(-1, 1))
             proba_drift = False
         else:
             # Flag for computing drift on the probabilities rather than the predicted labels
